requirements.txt
django-allauth
beautifulsoup4
requests
Pillow
